reworked to breakpoint can be set in any class at any time

+++ src/smthelusive/debyter/domain/AClass.java
@@ -1,6 +1,6 @@
 package smthelusive.debyter.domain;
 
-public record AClass(byte refTypeTag, long typeID, String signature, int status) {
+public record AClass(long refTypeTag, long typeID, String signature, int status) {
     @Override
     public String toString() {
         return "Clazz{" +
